pyproject.toml
[project]
name = "python-image-search"
version = "0.1.0"
description = ""
authors = [
    {name = "quangnguyen"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llm (>=0.25,<0.26)",
    "streamlit (>=1.45.1,<2.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "pillow-heif (>=0.22.0,<0.23.0)",
    "mediapipe (>=0.10.21,<0.11.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "duckduckgo-search (>=8.0.2,<9.0.0)",
    "clip @ git+https://github.com/openai/CLIP.git",
    "onnx (>=1.18.0,<2.0.0)",
    "onnxruntime (>=1.22.0,<2.0.0)",
    "torchvision (==0.16.0)",
    "torch (==2.1.0)",
    "mobileclip @ git+https://github.com/quangnd2203/ml-mobileclip.git@HEAD",

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false
